+++ JOB_PORTAL/JobPortal/src/main/java/com/jobportal/JobPortal/repository/UserRepository.java
@@ -7,12 +7,12 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.jobportal.JobPortal.model.User;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 	
 	@Query("select u from User u where u.username=?1")
 	User getByUsername(String username);
 		
-	Optional<User> getUserById(Long id);	
+	Optional<User> getUserById(int id);	
 	
    // Checks if an username is already registered
     boolean existsByUsername(String username);
