+++ JOB_PORTAL/JobPortal/src/main/java/com/jobportal/JobPortal/service/UserService.java
@@ -34,7 +34,7 @@ public class UserService {
 	}
  	
 //------------------------------------ Get user by ID ------------------------------
-    public User getUserById(Long userId) {
+    public User getUserById(int userId) {
         return userRepository.findById(userId)
             .orElseThrow(() -> new ResourceNotFoundException("User not found with ID: " + userId));
     }
